LHB-40 Add basic screens for nested Shell Route navigation
This facilitates a persistent bottom navigation bar for the app with branches for Home, Find, Favorites, and Settings.

These will be used by the GoRouter `StatefulShellRoute` and `ShellBranch`es. These are bare bones pages that allow for the nav bar to be built and demo the different areas of the applicaiton.

<!-- ps-id: fad814d0-32e9-4acf-980a-0087112139e0 -->