LHB-41 Change white native launch screen to black
This facilitates a clean transition from the native launch screen to the app Splash Screen.

Previously the app would display a white screen before the first frame is drawn. This makes the native launch screen black (this has not yet been tested on Android). This also updates the Splash Screen to make the background black, instead of the near-black from the App Theme Colors, for a smooth transition.

<!-- ps-id: 45fc3420-faeb-43f9-8c6c-d609d4167959 -->