LHB-44 Add AsyncValue extension to show snackbar on error
This facilitates displaying a snackbar if an error occurs on an AsyncValue.

For example, if categories cannot be retrieved when watching the categories provider, the async value can be listened to and the snackbar displayed to the user. This adds the AsyncValue extension and adds a Snackbar Theme so all snackbars are floating according to Material.

<!-- ps-id: 49c2a3b7-d380-4195-91f9-ee64a53b76ff -->