LHB-44 Add resizable White Bird Clinic logo widget
This toggles the logo between light and dark modes, and allows for the widget to be sized as needed. This will be used on the Home screen.

For now, this changes the color to black when the theme is in light mode, but we will want to swap this out for a true black logo in the future in order to maintain the fine details.

<!-- ps-id: 528b02fa-4431-49d0-81a6-4a1f3eb8b90a -->