Add the Services Screen
This facilitates retrieving and displaying the services by category.

This will become a new route within the Home tab of the app. The Home screen will have a card tile that routes to this screen so users can view services by the category. All categories are retrieved when the screen loads, and each tap on a category will retrieve and display the associated services.

<!-- ps-id: 0ed26840-ac51-4717-b847-62ab80612fb6 -->