Move PowerSync and Supabase directories into shared
This follows feature sliced design by moving shared code into the `/shared` directory, where it can be used by the layers above it.

This is necessary since a layer's segments should not be able to access other sgements within the layer, which would mean PowerSync and Supabase should not be in the `entities` directory.

<!-- ps-id: bab689d7-a5a5-4aaa-9fa8-de3676b6892c -->