Add category and service models to shared directory
This facilitates FSD by organizing shared models within the shared directory.

This creates the category and service models in the `/shared/models` directory, which share some components, such as the id typedef. These models are the closest definition to the datastore. All other enhancements to the model, such as business logic, will be in the `/entities` directory.

<!-- ps-id: 9a50dff1-b2f7-4930-b7a1-71a3b893a3c4 -->