Cleanup `simple_routes` migration after upgrade
This cleans up the workaround for overrides for the `goPath` getters in the route definitions. This also uses the `.fullPath()` for the `initialLocation` as recommended.

The initial overrides were implemented to facilitate the failure to build in another environment, where the app could not run due to a `Stack Overflow` error.

<!-- ps-id: dffd26fa-bbd2-42bb-875c-fbf3b81f6896 -->