LHB-57 Add Secondary Button with outline
This facilitates consistency across secondary action buttons across the app.

This defines a reusable component for a secondary call to action button which bordered using the `onSurface` color of the theme at the time. This also updates the BorderedContainer to use the `onSurface` color instead of `white` so it is versatile for any theme.

<!-- ps-id: cfb918a9-74e6-41ab-b3f2-3544e3dbaecf -->