LHB-83 Implement Search Bar on Providers Screen
This facilitates the filtering of providers by name or keyword in the description.

This adds the Providers Search Bar to the Providers Screen, and passes in a ValueNotifier, which updates the ProvidersList. If the Provider matches a search result, it will be displayed. If nos earch is provided, then all items are displayed.

This allows for fast search filtering, but the search is not yet exhaustive. For example, it does not search against the provider's services or categories, nor locations. This can be modified to be more exhaustive after release 1.0.0.

<!-- ps-id: c64bd85f-bde1-49a1-8ec8-1213ad2ddfaa -->