LHB-99 Add Emergency Crisis Lines list tile to Home screen
This facilitates accessing the emergency crisis lines from the home page when in need.

This routes to the Emergency Crisis Lines screen, which displays a list of searchable providers within the Crisis Lines category.

[changelog]
added: Emergency Crisis Lines list tile to Home screen to easily access providers when experiencing crisis

<!-- ps-id: cef447eb-d700-46bf-bc70-586000e97d9f -->