LHB-57 Add Schedule model and helper functions for formatting
This facilitates the retrieval and display of location schedules within the app, including hours of operation.

This defines the Schedule model and adds helper functions for formatting the schedule and determining whether the schedule is active. There are additional utilities to get the ordinal suffix of a numeric day, as well as to convert text to datetime or text to time of day.

<!-- ps-id: 6147dee8-229f-4620-a149-87f3789a0625 -->