LHB-49 Add global redirect for hard updates
This ensures users are aware when a hard update is required for the application whenever navigating the app.

This is to ensure all users are captured, particuarly those who never quit their applications. If a user does not quit and launch the app, the Splash screen won't catch them with the redirect, but the global redirect will.

<!-- ps-id: 1d06286a-70ea-458a-a62a-0fc8fce1d888 -->