LHB-104 Add locale to User Preferences schema, model, provider
This facilitates persistence and retrieval of the user's language preference of the application.

This supports the functionality of changing the app locale from English to Spanish and vice versa. This adds the column to the PowerSync schema, updates the model to include the property and parsing, and adds a factory function to build a User Preferences object with the default values.

Additionally, this includes the provider notifier method for updating the locale.

<!-- ps-id: 99f0da03-fce4-4f16-9ada-e25765d93fbf -->