LHB-104 Add L10n Toggle and toggle English to Spanish
This facilitates toggling the app to English or Spanish from the Setting s Screen.

This adds the L10n Toggle, adds it to the Settings Screen, and updates the app to listen for the User Preference. This also renames the `toggle_theme` directory to `toggle` to house all features that can toggle.

<!-- ps-id: 8f11d0cc-5b89-42f2-a798-e1f62d77242e -->