LHB-111: Setup Loggy Integration
Allow a developer to properly log and understand issues that
may arise. Loggy is the defecto logging solution which allows
a developer to easily customize how logs work.

This allows logs to be sent to Sentry by using logger.sevre
and Sentry breadcrumbs get added when using logger.info. These
are the default settings on Sentry’s Logging Integration.